# backend/openapi.yaml
openapi: 3.1.0
info:
  title: THUIE API
  version: 1.0.0-draft
  description: >-
    REST + WebSocket contract for the THUIE campus alumni info-exchange
    platform. Design skeleton - mirrors BACKEND.md and the PostgreSQL schema in
    PROJECT.md 6. Server implementation deferred to the build phase.
  license: { name: Proprietary }
servers:
  - url: https://api.thuie.example.com/api/v1
    description: Production (placeholder)
  - url: http://localhost:3000/api/v1
    description: Local development

# OpenAPI has no native WebSocket binding, so the realtime surface (BACKEND.md §9)
# is documented here as a vendor extension. Socket.IO server, namespace
# `/realtime`, authenticated with the same access JWT supplied via handshake
# `auth.token`, an `Authorization: Bearer` header, or a `?token=` query.
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post: { tags: [Auth], security: [], operationId: login, summary: Authenticate with handle + password,
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } },
      responses: { '200': { description: Token pair + user, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /auth/refresh:
    post: { tags: [Auth], security: [], operationId: refreshToken, summary: Rotate refresh token,
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { refresh_token: { type: string } } } } } },
      responses: { '200': { description: New token pair, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /auth/logout:
    post: { tags: [Auth], operationId: logout, summary: Revoke presented refresh token (or the whole device family),
      requestBody: { content: { application/json: { schema: { type: object, properties: { refresh_token: { type: string }, all_devices: { type: boolean } } } } } },
      responses: { '200': { description: Logged out }, default: { $ref: '#/components/responses/Error' } } }
  /auth/register:
    post: { tags: [Auth], security: [], operationId: register, summary: Student self-signup,
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ],
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } } },
      responses: { '201': { description: Created user + tokens, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /auth/password:
    patch: { tags: [Auth], operationId: changePassword, summary: Change password (invalidates older tokens),
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [current_password, new_password], properties: { current_password: { type: string }, new_password: { type: string } } } } } },
      responses: { '200': { description: Updated }, default: { $ref: '#/components/responses/Error' } } }
  # ---- Users / me ----
  /me:
    get: { tags: [Users], operationId: getMe, summary: Current user, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Users], operationId: updateMe, summary: Update profile fields, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/UserUpdate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Users], operationId: deleteMe, summary: Soft delete + PII scrub, responses: { '204': { description: deleted }, default: { $ref: '#/components/responses/Error' } } }
...
  /faculty:
    get: { tags: [Faculty], operationId: listFaculty, summary: Faculty directory, parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: department, schema: { type: string } }, { in: query, name: q, schema: { type: string } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/FacultyPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /faculty/{id}:
    get: { tags: [Faculty], operationId: getFaculty, summary: Faculty member, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/FacultyMember' } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Forum ----
...
  /connections:
    get: { tags: [Connections], operationId: listConnections, summary: requests | mine, parameters: [ { in: query, name: box, schema: { type: string, enum: [requests, mine] } } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Connection' } } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Connections], operationId: createConnection, summary: Send connection request, parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { to_user_id: { type: string }, message: { type: string } } } } } }, responses: { '201': { description: pending, content: { application/json: { schema: { $ref: '#/components/schemas/Connection' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /connections/{id}:
    patch: { tags: [Connections], operationId: respondConnection, summary: accept / decline / revoke, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { status: { $ref: '#/components/schemas/ConnectionStatus' } } } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Connection' } } } }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Events ----
  /events:
    get: { tags: [Events], operationId: listEvents, summary: Published events, parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: type, schema: { $ref: '#/components/schemas/EventType' } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/EventPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Events], operationId: createEvent, summary: Admin create event, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/EventCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/CampusEvent' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /events/{id}:
    get: { tags: [Events], operationId: getEvent, summary: Event detail, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/CampusEvent' } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Events], operationId: updateEvent, summary: Admin edit / status, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/EventCreate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/CampusEvent' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /events/{id}/register:
    post: { tags: [Events], operationId: registerEvent, summary: Register (capacity lock), parameters: [ { $ref: '#/components/parameters/Id' }, { $ref: '#/components/parameters/IdempotencyKey' } ], responses: { '201': { description: registered, content: { application/json: { schema: { $ref: '#/components/schemas/EventRegistration' } } } }, '409': { description: capacity reached, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Events], operationId: cancelRegistration, summary: Cancel own registration, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: cancelled }, default: { $ref: '#/components/responses/Error' } } }
  /events/{id}/ticket:
    get: { tags: [Events], operationId: getEventTicket, summary: Signed single-use ticket, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /events/{id}/check-in:
    post: { tags: [Events], operationId: checkInEvent, summary: Organizer verify scanned ticket, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, required: [ticket], properties: { ticket: { type: string } } } } } }, responses: { '200': { description: attended, content: { application/json: { schema: { $ref: '#/components/schemas/EventRegistration' } } } }, '409': { description: replay/expired, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Mentorship ----
...
  /admin/faculty:
    get: { tags: [Admin], operationId: adminListFaculty, summary: Faculty admin table (offset + q/sort/order), parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: q, schema: { type: string } }, { in: query, name: sort, schema: { type: string, enum: [name, title] } }, { in: query, name: order, schema: { type: string, enum: [asc, desc] } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/FacultyPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Admin], operationId: adminCreateFaculty, summary: Create faculty member, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/FacultyInput' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/FacultyMember' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/faculty/{id}:
    patch: { tags: [Admin], operationId: adminUpdateFaculty, summary: Edit faculty member, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/FacultyInput' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/FacultyMember' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Admin], operationId: adminDeleteFaculty, summary: Delete faculty member, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: deleted }, default: { $ref: '#/components/responses/Error' } } }
...
  /auth/forgot-password:
    post: { tags: [Auth], security: [], operationId: forgotPassword, summary: Send password-reset OTP/link,
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ForgotPasswordRequest' } } } },
      responses: { '202': { description: accepted (always) }, default: { $ref: '#/components/responses/Error' } } }
  /auth/reset-password:
    post: { tags: [Auth], security: [], operationId: resetPassword, summary: Redeem reset code (revokes all sessions),
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ResetPasswordRequest' } } } },
      responses: { '200': { description: reset done }, default: { $ref: '#/components/responses/Error' } } }
  /auth/verify:
    post: { tags: [Auth], security: [], operationId: verifyOtp, summary: Confirm an OTP,
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OtpVerifyRequest' } } } },
      responses: { '200': { description: verified, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /auth/resend:
    post: { tags: [Auth], security: [], operationId: resendOtp, summary: Re-issue an OTP,
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [identifier, purpose], properties: { identifier: { type: string }, purpose: { $ref: '#/components/schemas/OtpPurpose' } } } } } },
      responses: { '200': { description: sent }, default: { $ref: '#/components/responses/Error' } } }
  /auth/otp:
    post: { tags: [Auth], security: [], operationId: issueOtp, summary: Issue/refresh an OTP for an identifier (query form, convenience for the client's send-code-first step),
      parameters: [ { in: query, name: identifier, required: true, schema: { type: string } }, { in: query, name: purpose, schema: { $ref: '#/components/schemas/OtpPurpose' } } ],
      responses: { '200': { description: sent }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Users: sessions / my content / blocks / prefs ----
...
  /me/blocks:
    get: { tags: [Connections], operationId: listBlocks, summary: Blocked users, responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Block' } } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Connections], operationId: blockUser, summary: Block a user, requestBody: { content: { application/json: { schema: { type: object, properties: { user_id: { type: string } } } } } }, responses: { '201': { description: blocked }, default: { $ref: '#/components/responses/Error' } } }
  /me/blocks/{userId}:
    delete: { tags: [Connections], operationId: unblockUser, summary: Unblock, parameters: [ { $ref: '#/components/parameters/UserId' } ], responses: { '204': { description: unblocked }, default: { $ref: '#/components/responses/Error' } } }
...
  /me/event-registrations:
    get: { tags: [Events], operationId: listMyRegistrations, summary: My registrations + tickets, parameters: [ { in: query, name: upcoming, schema: { type: boolean } } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/EventRegistration' } } } } }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Search suggest ----
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
    Id: { in: path, name: id, required: true, schema: { type: string } }
    UserId: { in: path, name: userId, required: true, schema: { type: string } }
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
...
    IdempotencyKey: { in: header, name: Idempotency-Key, schema: { type: string, format: uuid } }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found (or hidden by visibility)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    Role: { type: string, enum: [student, graduate, admin, admin_super] }
    UserStatus: { type: string, enum: [unverified, active, posting_restricted, banned, deleted] }
    OtpPurpose: { type: string, enum: [register_verify, password_reset] }
...
    Visibility: { type: string, enum: [student_only, all, admin_only] }
...
    EventType: { type: string, enum: [recruitment, lecture, sharing, ceremony, sports, other] }
    ConnectionStatus: { type: string, enum: [pending, accepted, declined, revoked] }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine code }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }
        errors: { type: array, items: { type: object, properties: { field: { type: string }, message: { type: string } } } }
...
    OffsetMeta: { type: object, properties: { page: { type: integer }, limit: { type: integer }, total: { type: integer }, totalPages: { type: integer } } }
...
    LoginRequest: { type: object, required: [handle, password], properties: { handle: { type: string, description: student_id / phone / email }, password: { type: string }, device_label: { type: string } } }
    RegisterRequest: { type: object, required: [name, role, password], properties: { name: { type: string }, role: { type: string, enum: [student, graduate] }, student_id: { type: string }, phone: { type: string }, email: { type: string }, password: { type: string }, department: { type: string, description: department code or id }, grade_year: { type: string }, otp_code: { type: string }, invite_code: { type: string } } }
    TokenPair: { type: object, properties: { access_token: { type: string }, refresh_token: { type: string }, expires_in: { type: integer }, user: { $ref: '#/components/schemas/User' } } }
...
    Department: { type: object, properties: { id: { type: string }, code: { type: string }, name_zh: { type: string }, name_en: { type: string }, faculty: { type: [string, 'null'] } } }
    User: { type: object, properties: { id: { type: string }, name: { type: string }, role: { $ref: '#/components/schemas/Role' }, status: { $ref: '#/components/schemas/UserStatus' }, student_id: { type: [string, 'null'] }, phone: { type: [string, 'null'] }, email: { type: [string, 'null'] }, department: { $ref: '#/components/schemas/Department' }, grade_year: { type: [string, 'null'] }, graduation_year: { type: [string, 'null'] }, avatar_url: { type: [string, 'null'] }, bio: { type: [string, 'null'] }, wechat: { type: [string, 'null'], description: GET /me only }, whatsapp: { type: [string, 'null'], description: GET /me only }, linkedin: { type: [string, 'null'], description: GET /me only }, wechat_visibility: { $ref: '#/components/schemas/Visibility' }, whatsapp_visibility: { $ref: '#/components/schemas/Visibility' }, linkedin_visibility: { $ref: '#/components/schemas/Visibility' }, email_visibility: { $ref: '#/components/schemas/Visibility' }, phone_visibility: { $ref: '#/components/schemas/Visibility' }, is_online: { type: boolean }, created_at: { type: string, format: date-time } } }
    UserUpdate: { type: object, properties: { name: { type: string }, bio: { type: string }, avatar_media_id: { type: string }, grade_year: { type: string }, wechat: { type: string }, whatsapp: { type: string }, linkedin: { type: string }, wechat_visibility: { $ref: '#/components/schemas/Visibility' }, whatsapp_visibility: { $ref: '#/components/schemas/Visibility' }, linkedin_visibility: { $ref: '#/components/schemas/Visibility' }, email_visibility: { $ref: '#/components/schemas/Visibility' }, phone_visibility: { $ref: '#/components/schemas/Visibility' } } }
...
    MediaObject: { type: object, properties: { id: { type: string }, kind: { type: string, enum: [image, attachment, avatar] }, url: { type: string }, file_name: { type: string }, mime_type: { type: string }, size_bytes: { type: integer }, width: { type: [integer, 'null'] }, height: { type: [integer, 'null'] }, thumbnail_key: { type: [string, 'null'] } } }
...
    FacultyMember: { type: object, properties: { id: { type: string }, name: { type: string }, department: { $ref: '#/components/schemas/Department' }, title: { type: string }, research_area: { type: string }, email: { type: [string, 'null'] }, phone: { type: [string, 'null'] }, bio: { type: [string, 'null'] }, avatar_url: { type: [string, 'null'] } } }
    FacultyInput: { type: object, required: [name], properties: { name: { type: string, maxLength: 100 }, department: { type: string, description: department code or id }, title: { type: string, maxLength: 120 }, research_area: { type: string, maxLength: 240 }, email: { type: string }, phone: { type: string, maxLength: 40 }, bio: { type: string, maxLength: 2000 }, avatar_media_id: { type: string }, avatar_url: { type: string }, address: { type: string, maxLength: 240 }, homepage: { type: string, maxLength: 500 } } }
...
    Connection: { type: object, properties: { id: { type: string }, from_user: { $ref: '#/components/schemas/User' }, to_user: { $ref: '#/components/schemas/User' }, status: { $ref: '#/components/schemas/ConnectionStatus' }, message: { type: [string, 'null'] }, responded_at: { type: [string, 'null'], format: date-time }, created_at: { type: string, format: date-time } } }
    CampusEvent: { type: object, properties: { id: { type: string }, title: { type: string }, description: { type: string }, starts_at: { type: string, format: date-time }, ends_at: { type: [string, 'null'], format: date-time }, location: { type: string }, organizer: { type: string }, type: { $ref: '#/components/schemas/EventType' }, capacity: { type: integer }, registered_count: { type: integer }, status: { type: string, enum: [draft, published, cancelled, archived] }, cover: { oneOf: [ { $ref: '#/components/schemas/MediaObject' }, { type: 'null' } ] } } }
    EventCreate: { type: object, required: [title, starts_at], properties: { title: { type: string }, description: { type: string }, starts_at: { type: string, format: date-time }, ends_at: { type: string, format: date-time }, location: { type: string }, organizer: { type: string }, type: { $ref: '#/components/schemas/EventType' }, capacity: { type: integer, minimum: 0 }, status: { type: string, enum: [draft, published, cancelled, archived] }, cover_media_id: { type: string } } }
    EventRegistration: { type: object, properties: { event_id: { type: string }, user_id: { type: string }, registered_at: { type: string, format: date-time }, status: { type: string, enum: [registered, attended, cancelled] } } }
    Ticket: { type: object, properties: { token: { type: string, description: Signed, short-lived, single-use }, expires_at: { type: string, format: date-time } } }
...
    ForgotPasswordRequest: { type: object, required: [identifier], properties: { identifier: { type: string } } }
    ResetPasswordRequest: { type: object, required: [identifier, code, new_password], properties: { identifier: { type: string }, code: { type: string }, new_password: { type: string } } }
    OtpVerifyRequest: { type: object, required: [identifier, purpose, code], properties: { identifier: { type: string, description: the phone/email that received the code }, purpose: { $ref: '#/components/schemas/OtpPurpose' }, code: { type: string } } }
...
    Block: { type: object, properties: { blocked_id: { type: string }, user: { $ref: '#/components/schemas/User' }, created_at: { type: string, format: date-time } } }
...
    FacultyPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/FacultyMember' } }, meta: { $ref: '#/components/schemas/OffsetMeta' } } }
...
    EventPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/CampusEvent' } }, meta: { $ref: '#/components/schemas/OffsetMeta' } } }
...